Rename api/lib → api/_lib so Vercel skips helper files
Vercel deploys all files under api/ as serverless functions. The lib/
directory contained helper modules (orchestrator, agent-loader, router,
cors) that were incorrectly being deployed as standalone endpoints,
causing deployment failures.

The underscore prefix (_lib/) is Vercel's convention for directories
that should be bundled as dependencies but NOT deployed as functions.
