+++ api/agents.js
@@ -1,10 +1,10 @@
-import { cors } from './lib/cors.js'
+import { cors } from './_lib/cors.js'
 
 export default async function handler(req, res) {
   if (cors(req, res)) return
 
   try {
-    const { getAgentRegistry } = await import('./lib/agent-loader.js')
+    const { getAgentRegistry } = await import('./_lib/agent-loader.js')
     const registry = await getAgentRegistry()
     res.status(200).json({ agents: registry })
   } catch (err) {
