+++ vercel.json
@@ -5,7 +5,7 @@
   "functions": {
     "api/**/*.js": {
       "maxDuration": 60,
-      "includeFiles": ["agents/**", "agents/workflows/**"]
+      "includeFiles": ["agents/**", "agents/workflows/**", "api/_lib/**"]
     }
   }
 }
